+++ pkg/sqsHandlers/handlers.go
@@ -1,4 +1,4 @@
-package handlers
+package sqsHandlers
 
 import "github.com/aws/aws-sdk-go/service/sqs"
