+++ pkg/sqsHandlers/inspectorHandlers.go
@@ -1,4 +1,4 @@
-package handlers
+package sqsHandlers
 
 import (
 	"bytes"
@@ -27,11 +27,11 @@ func (h *inspectorHandler) HandleMessage(msg *sqs.Message) error {
 	b := []byte(*msg.Body)
 	f, err := inspectorHelper.ReadFindingFromBytes(b)
 	if err != nil {
-		return fmt.Errorf("reading finding: %v", err)
+		return fmt.Errorf("reading inspector finding: %v", err)
 	}
 	err = h.handleFinding(*f)
 	if err != nil {
-		logger.Error(err, "handling finding")
+		logger.Error(err, "handling inspector finding")
 	}
 	return nil
 }
@@ -40,21 +40,16 @@ func (h *inspectorHandler) handleFinding(f inspector2.Finding) error {
 	h.manager.InsertInspectorFinding(f)
 	b, err := json.Marshal(h.manager.Document())
 	if err != nil {
-		return fmt.Errorf("marshaling json: %v", err)
+		return fmt.Errorf("marshaling inspector finding to json: %v", err)
 	}
 	buf := &bytes.Buffer{}
 	err = json.Indent(buf, b, "", "\t")
 	if err != nil {
-		return fmt.Errorf("formating json: %v", err)
+		return fmt.Errorf("formating inspector finding json: %v", err)
 	}
 	_, err = os.Stdout.Write(buf.Bytes())
 	if err != nil {
-		return fmt.Errorf("writing to stdout: %v", err)
+		return fmt.Errorf("writing inspector finding json to stdout: %v", err)
 	}
-	return nil
-}
-
-func (h *inspectorHandler) DeleteMessage(msg *sqs.Message) error {
-	fmt.Println("this is where we'd delete the message")
 	return nil
 }
